decompose steps to be reusable
